refs #2 - Corrections code smells

+++ src/utilityPivot.cpp
@@ -20,7 +20,7 @@
 */
 long UtilityPivot::toTimestamp(long secondSinceEpoch, long fractionOfSecond) {
     long timestamp = 0;
-    long msPart = (long)(round((fractionOfSecond * 1000L) / 16777216.0));
+    long msPart = (long)round((double)(fractionOfSecond * 1000) / 16777216.0);
     timestamp = (secondSinceEpoch * 1000L) + msPart;
     return timestamp;
 }
